Do not reap pods; allow failed pods to remain for log inspection.
When the pod reaper runs it removes all pods where `pod.status.phase == "Succeeded"`. Apparently this includes pods that were OOMKilled. The Kubernetes docs state "When you delete the job using kubectl, all the pods it created are deleted too." So we no longer need to reap pods and if a pod fails we would like to have it around to inspect the logs and events.